refactor: split e2b provider boundaries (#74)
