+++ internal/runtime/e2bruntime/runtime.go
@@ -1,12 +1,8 @@
 package e2bruntime
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
-	pathpkg "path"
 	"strings"
 	"time"
 
@@ -15,16 +11,6 @@ import (
 	skillsapi "github.com/superduck-ai/open-managed-agents/internal/skills"
 
 	e2b "github.com/superduck-ai/e2b-go-sdk"
-)
-
-const (
-	sandboxUserDataVolumeName = "user-data"
-	sandboxUserDataMountPath  = "/mnt/user-data"
-	SandboxSkillsMountPath    = "/mnt/skills"
-	SkillMountMetadataKey     = "managed_agent_skills_mount"
-	skillMountManifestPath    = "manifest.json"
-	skillMountReadyPath       = ".ready"
-	skillMountVolumePrefix    = "managed-agent-skills-"
 )
 
 type Resolution struct {
@@ -40,13 +26,6 @@ type Sandbox struct {
 	ID string
 }
 
-type SkillMount struct {
-	MountPath      string                         `json:"mount_path"`
-	VolumeName     string                         `json:"volume_name"`
-	ManifestSHA256 string                         `json:"manifest_sha256"`
-	Skills         []skillsapi.MountManifestSkill `json:"skills,omitempty"`
-}
-
 type Provider interface {
 	Create(ctx context.Context, env db.Environment, work *db.EnvironmentWork, resolution Resolution) (Sandbox, error)
 	Kill(ctx context.Context, sandboxID string) error
@@ -60,11 +39,15 @@ type SkillMountPreparer interface {
 }
 
 type E2BProvider struct {
-	cfg config.Config
+	cfg       config.Config
+	sandboxes sandboxOperations
 }
 
 func NewProvider(cfg config.Config) *E2BProvider {
-	return &E2BProvider{cfg: cfg}
+	return &E2BProvider{
+		cfg:       cfg,
+		sandboxes: newE2BSandboxOperations(cfg),
+	}
 }
 
 func (p *E2BProvider) Resolve(env db.Environment, work *db.EnvironmentWork) (Resolution, error) {
@@ -94,454 +77,4 @@ func (p *E2BProvider) Resolve(env db.Environment, work *db.EnvironmentWork) (Res
 	resolved.AllowInternetAccess = allowInternet
 	resolved.Network = network
 	return resolved, nil
-}
-
-func (p *E2BProvider) Create(ctx context.Context, env db.Environment, work *db.EnvironmentWork, resolved Resolution) (Sandbox, error) {
-	if strings.TrimSpace(p.cfg.E2BAPIKey) == "" && !p.cfg.E2BDebug {
-		return Sandbox{}, errors.New("E2B_API_KEY is required to create a sandbox")
-	}
-	if strings.TrimSpace(resolved.Template) == "" {
-		var err error
-		resolved, err = p.Resolve(env, work)
-		if err != nil {
-			return Sandbox{}, err
-		}
-	}
-	timeoutMs := int(resolved.Timeout / time.Millisecond)
-	if timeoutMs <= 0 {
-		timeoutMs = int((5 * time.Minute) / time.Millisecond)
-	}
-	requestTimeoutMs := int(p.cfg.E2BRequestTimeout / time.Millisecond)
-	debug := p.cfg.E2BDebug
-	allowInternet := resolved.AllowInternetAccess
-	opts := &e2b.SandboxOpts{
-		ConnectionOpts: e2b.ConnectionOpts{
-			ApiKey:           p.cfg.E2BAPIKey,
-			AccessToken:      p.cfg.E2BAccessToken,
-			Domain:           p.cfg.E2BDomain,
-			ApiUrl:           p.cfg.E2BAPIURL,
-			SandboxUrl:       p.cfg.E2BSandboxURL,
-			Debug:            &debug,
-			RequestTimeoutMs: &requestTimeoutMs,
-		},
-		Metadata:            resolved.Metadata,
-		Envs:                resolved.Envs,
-		TimeoutMs:           &timeoutMs,
-		AllowInternetAccess: &allowInternet,
-		Network:             resolved.Network,
-	}
-	if volumeMounts := p.sandboxVolumeMounts(work); len(volumeMounts) > 0 {
-		opts.VolumeMounts = volumeMounts
-	}
-	sandbox, err := e2b.Create(ctx, resolved.Template, opts)
-	if err != nil {
-		return Sandbox{}, err
-	}
-	return Sandbox{ID: sandbox.SandboxID}, nil
-}
-
-func (p *E2BProvider) Kill(ctx context.Context, sandboxID string) error {
-	if strings.TrimSpace(sandboxID) == "" {
-		return nil
-	}
-	sandbox, err := p.connect(ctx, sandboxID)
-	if err != nil {
-		return err
-	}
-	return sandbox.Kill(ctx, nil)
-}
-
-func (p *E2BProvider) WriteFile(ctx context.Context, sandboxID string, filePath string, data []byte) error {
-	if strings.TrimSpace(sandboxID) == "" {
-		return errors.New("sandbox id is required")
-	}
-	if strings.TrimSpace(filePath) == "" {
-		return errors.New("sandbox file path is required")
-	}
-	sandbox, err := p.connect(ctx, sandboxID)
-	if err != nil {
-		return err
-	}
-	if dir := pathpkg.Dir(filePath); dir != "." && dir != "/" {
-		if _, err := sandbox.Commands.Run(ctx, "mkdir -p "+shellQuote(dir), nil); err != nil {
-			return err
-		}
-	}
-	_, err = sandbox.Files.Write(ctx, filePath, bytes.NewReader(data), nil)
-	return err
-}
-
-func (p *E2BProvider) RunCommand(ctx context.Context, sandboxID string, command string) error {
-	if strings.TrimSpace(sandboxID) == "" {
-		return errors.New("sandbox id is required")
-	}
-	if strings.TrimSpace(command) == "" {
-		return errors.New("sandbox command is required")
-	}
-	sandbox, err := p.connect(ctx, sandboxID)
-	if err != nil {
-		return err
-	}
-	timeoutMs := int(p.cfg.E2BRequestTimeout / time.Millisecond)
-	if timeoutMs <= 0 {
-		timeoutMs = int((60 * time.Second) / time.Millisecond)
-	}
-	execution, err := sandbox.Commands.Run(ctx, command, &e2b.CommandStartOpts{TimeoutMs: &timeoutMs})
-	if err != nil {
-		var exitErr *e2b.CommandExitError
-		if errors.As(err, &exitErr) {
-			return fmt.Errorf("sandbox command exited with code %d: %s stdout=%q stderr=%q", exitErr.ExitCode, strings.TrimSpace(exitErr.Message), truncateCommandOutput(exitErr.Stdout), truncateCommandOutput(exitErr.Stderr))
-		}
-		return err
-	}
-	result, ok := execution.(*e2b.CommandResult)
-	if !ok {
-		return fmt.Errorf("sandbox command execution type = %T, want *e2b.CommandResult", execution)
-	}
-	if result.ExitCode != 0 {
-		return fmt.Errorf("sandbox command exited with code %d: %s stdout=%q stderr=%q", result.ExitCode, strings.TrimSpace(result.Error), truncateCommandOutput(result.Stdout), truncateCommandOutput(result.Stderr))
-	}
-	return nil
-}
-
-func (p *E2BProvider) PrepareSkillMount(ctx context.Context, runtimeSkills []skillsapi.RuntimeSkill) (*SkillMount, error) {
-	if len(runtimeSkills) == 0 {
-		return nil, nil
-	}
-	manifest, manifestBytes, manifestSHA256, err := skillsapi.BuildMountManifest(runtimeSkills)
-	if err != nil {
-		return nil, err
-	}
-	volumeName := skillMountVolumeName(manifestSHA256)
-	volume, created, err := p.connectOrCreateSkillVolume(ctx, volumeName)
-	if err != nil {
-		return nil, err
-	}
-	if !created {
-		ready, err := p.skillVolumeReady(ctx, volume, volumeName, manifestSHA256)
-		if err != nil {
-			return nil, err
-		}
-		if ready {
-			return &SkillMount{
-				MountPath:      SandboxSkillsMountPath,
-				VolumeName:     volumeName,
-				ManifestSHA256: manifestSHA256,
-				Skills:         manifest.Skills,
-			}, nil
-		}
-	}
-	if err := p.writeSkillVolume(ctx, volume, manifestBytes, manifestSHA256, manifest, runtimeSkills); err != nil {
-		return nil, err
-	}
-	return &SkillMount{
-		MountPath:      SandboxSkillsMountPath,
-		VolumeName:     volumeName,
-		ManifestSHA256: manifestSHA256,
-		Skills:         manifest.Skills,
-	}, nil
-}
-
-func (p *E2BProvider) skillVolumeReady(ctx context.Context, volume *e2b.Volume, volumeName string, manifestSHA256 string) (bool, error) {
-	exists, err := volume.Exists(ctx, skillMountReadyPath, p.volumeAPIOpts())
-	if err != nil {
-		return false, fmt.Errorf("check managed agent skill volume readiness %s: %w", volumeName, err)
-	}
-	if !exists {
-		return false, nil
-	}
-	value, err := volume.ReadFile(ctx, skillMountReadyPath, p.volumeReadOpts())
-	if err != nil {
-		return false, fmt.Errorf("read managed agent skill volume ready marker %s: %w", volumeName, err)
-	}
-	var marker string
-	switch typed := value.(type) {
-	case string:
-		marker = typed
-	case []byte:
-		marker = string(typed)
-	default:
-		return false, fmt.Errorf("managed agent skill volume ready marker %s has unsupported type %T", volumeName, value)
-	}
-	return strings.TrimSpace(marker) == strings.TrimSpace(manifestSHA256), nil
-}
-
-func (p *E2BProvider) connect(ctx context.Context, sandboxID string) (*e2b.Sandbox, error) {
-	requestTimeoutMs := int(p.cfg.E2BRequestTimeout / time.Millisecond)
-	debug := p.cfg.E2BDebug
-	sandbox, err := e2b.Connect(ctx, sandboxID, &e2b.SandboxConnectOpts{
-		ConnectionOpts: e2b.ConnectionOpts{
-			ApiKey:           p.cfg.E2BAPIKey,
-			AccessToken:      p.cfg.E2BAccessToken,
-			Domain:           p.cfg.E2BDomain,
-			ApiUrl:           p.cfg.E2BAPIURL,
-			SandboxUrl:       p.cfg.E2BSandboxURL,
-			Debug:            &debug,
-			RequestTimeoutMs: &requestTimeoutMs,
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	return sandbox, nil
-}
-
-func (p *E2BProvider) connectOrCreateSkillVolume(ctx context.Context, volumeName string) (*e2b.Volume, bool, error) {
-	opts := p.volumeConnectionOpts()
-	volumes, err := e2b.ListVolumes(ctx, opts)
-	if err != nil {
-		return nil, false, fmt.Errorf("list E2B volumes for managed agent skills: %w", err)
-	}
-	for _, volume := range volumes {
-		if volume.Name != volumeName {
-			continue
-		}
-		connected, err := e2b.ConnectVolume(ctx, volume.VolumeID, opts)
-		if err != nil {
-			return nil, false, fmt.Errorf("connect managed agent skill volume %s: %w", volumeName, err)
-		}
-		return connected, false, nil
-	}
-	created, err := e2b.CreateVolume(ctx, volumeName, opts)
-	if err == nil {
-		return created, true, nil
-	}
-	volumes, listErr := e2b.ListVolumes(ctx, opts)
-	if listErr == nil {
-		for _, volume := range volumes {
-			if volume.Name != volumeName {
-				continue
-			}
-			connected, connectErr := e2b.ConnectVolume(ctx, volume.VolumeID, opts)
-			if connectErr != nil {
-				return nil, false, fmt.Errorf("connect concurrently created managed agent skill volume %s: %w", volumeName, connectErr)
-			}
-			return connected, false, nil
-		}
-	}
-	return nil, false, fmt.Errorf("create managed agent skill volume %s: %w", volumeName, err)
-}
-
-func (p *E2BProvider) writeSkillVolume(ctx context.Context, volume *e2b.Volume, manifestBytes []byte, manifestSHA256 string, manifest skillsapi.MountManifest, runtimeSkills []skillsapi.RuntimeSkill) error {
-	skillsByFilename := make(map[string]skillsapi.RuntimeSkill, len(runtimeSkills))
-	for _, skill := range runtimeSkills {
-		skillsByFilename[skillsapi.MountArchiveFilename(skill)] = skill
-	}
-	opts := p.volumeWriteOpts()
-	if _, err := volume.WriteFile(ctx, skillMountManifestPath, manifestBytes, opts); err != nil {
-		return fmt.Errorf("write managed agent skill manifest: %w", err)
-	}
-	for _, entry := range manifest.Skills {
-		skill, ok := skillsByFilename[entry.Filename]
-		if !ok {
-			return fmt.Errorf("managed agent skill archive %s is missing", entry.Filename)
-		}
-		archive, err := skill.LoadArchive(ctx)
-		if err != nil {
-			return err
-		}
-		if len(archive) == 0 {
-			return fmt.Errorf("managed agent skill archive %s is empty", entry.Filename)
-		}
-		if _, err := volume.WriteFile(ctx, entry.Filename, archive, opts); err != nil {
-			return fmt.Errorf("write managed agent skill archive %s: %w", entry.Filename, err)
-		}
-	}
-	if _, err := volume.WriteFile(ctx, skillMountReadyPath, []byte(manifestSHA256+"\n"), opts); err != nil {
-		return fmt.Errorf("write managed agent skill volume ready marker: %w", err)
-	}
-	return nil
-}
-
-func (p *E2BProvider) volumeConnectionOpts() *e2b.VolumeConnectionOpts {
-	requestTimeoutMs := int(p.cfg.E2BRequestTimeout / time.Millisecond)
-	debug := p.cfg.E2BDebug
-	return &e2b.VolumeConnectionOpts{
-		ApiKey:           p.cfg.E2BAPIKey,
-		AccessToken:      p.cfg.E2BAccessToken,
-		Domain:           p.cfg.E2BDomain,
-		ApiUrl:           p.cfg.E2BAPIURL,
-		SandboxUrl:       p.cfg.E2BSandboxURL,
-		Debug:            &debug,
-		RequestTimeoutMs: &requestTimeoutMs,
-	}
-}
-
-func (p *E2BProvider) volumeAPIOpts() *e2b.VolumeApiOpts {
-	requestTimeoutMs := int(p.cfg.E2BRequestTimeout / time.Millisecond)
-	debug := p.cfg.E2BDebug
-	return &e2b.VolumeApiOpts{
-		Domain:           p.cfg.E2BDomain,
-		Debug:            &debug,
-		ApiUrl:           p.cfg.E2BAPIURL,
-		RequestTimeoutMs: &requestTimeoutMs,
-	}
-}
-
-func (p *E2BProvider) volumeReadOpts() *e2b.VolumeReadOpts {
-	apiOpts := p.volumeAPIOpts()
-	return &e2b.VolumeReadOpts{
-		VolumeApiOpts: *apiOpts,
-	}
-}
-
-func (p *E2BProvider) volumeWriteOpts() *e2b.VolumeWriteOptions {
-	force := true
-	mode := 0o644
-	apiOpts := p.volumeAPIOpts()
-	return &e2b.VolumeWriteOptions{
-		VolumeMetadataOptions: e2b.VolumeMetadataOptions{Mode: &mode},
-		Force:                 &force,
-		Domain:                apiOpts.Domain,
-		Debug:                 apiOpts.Debug,
-		ApiUrl:                apiOpts.ApiUrl,
-		RequestTimeoutMs:      apiOpts.RequestTimeoutMs,
-	}
-}
-
-func resolveNetwork(raw json.RawMessage, mcpAllowedHosts []string) (*e2b.SandboxNetworkOpts, bool, error) {
-	var config struct {
-		Type       string `json:"type"`
-		Networking *struct {
-			Type                 string   `json:"type"`
-			AllowedHosts         []string `json:"allowed_hosts"`
-			AllowPackageManagers bool     `json:"allow_package_managers"`
-			AllowMCPServers      bool     `json:"allow_mcp_servers"`
-		} `json:"networking"`
-	}
-	if len(raw) == 0 {
-		return nil, true, nil
-	}
-	if err := json.Unmarshal(raw, &config); err != nil {
-		return nil, false, err
-	}
-	if config.Type != "cloud" || config.Networking == nil || config.Networking.Type == "unrestricted" {
-		return nil, true, nil
-	}
-	if config.Networking.Type != "limited" {
-		return nil, false, nil
-	}
-	hosts := append([]string(nil), config.Networking.AllowedHosts...)
-	if config.Networking.AllowPackageManagers {
-		hosts = append(hosts, packageManagerHosts()...)
-	}
-	if config.Networking.AllowMCPServers {
-		hosts = append(hosts, mcpAllowedHosts...)
-	}
-	return &e2b.SandboxNetworkOpts{AllowOut: uniqueStrings(hosts)}, false, nil
-}
-
-func mcpAllowedHostsFromWork(work *db.EnvironmentWork) []string {
-	if work == nil || len(work.Metadata) == 0 || strings.TrimSpace(string(work.Metadata)) == "null" {
-		return nil
-	}
-	var metadata map[string]any
-	if err := json.Unmarshal(work.Metadata, &metadata); err != nil {
-		return nil
-	}
-	values, ok := metadata["mcp_allowed_hosts"].([]any)
-	if !ok {
-		return nil
-	}
-	hosts := make([]string, 0, len(values))
-	for _, value := range values {
-		host, ok := value.(string)
-		if !ok {
-			continue
-		}
-		hosts = append(hosts, host)
-	}
-	return uniqueStrings(hosts)
-}
-
-func packageManagerHosts() []string {
-	return []string{
-		"archive.ubuntu.com",
-		"security.ubuntu.com",
-		"pypi.org",
-		"files.pythonhosted.org",
-		"registry.npmjs.org",
-		"proxy.golang.org",
-		"sum.golang.org",
-		"crates.io",
-		"index.crates.io",
-		"rubygems.org",
-	}
-}
-
-func uniqueStrings(values []string) []string {
-	seen := map[string]struct{}{}
-	out := make([]string, 0, len(values))
-	for _, value := range values {
-		value = strings.TrimSpace(value)
-		if value == "" {
-			continue
-		}
-		if _, ok := seen[value]; ok {
-			continue
-		}
-		seen[value] = struct{}{}
-		out = append(out, value)
-	}
-	return out
-}
-
-func shellQuote(value string) string {
-	if value == "" {
-		return "''"
-	}
-	return "'" + strings.ReplaceAll(value, "'", "'\"'\"'") + "'"
-}
-
-func truncateCommandOutput(value string) string {
-	value = strings.TrimSpace(value)
-	const limit = 2048
-	if len(value) <= limit {
-		return value
-	}
-	return value[:limit] + "...[truncated]"
-}
-
-func (p *E2BProvider) sandboxVolumeMounts(work *db.EnvironmentWork) map[string]any {
-	mounts := map[string]any{
-		sandboxUserDataMountPath: sandboxUserDataVolumeName,
-	}
-	if skillMount, ok := skillMountFromWork(work); ok {
-		mountPath := strings.TrimSpace(skillMount.MountPath)
-		if mountPath == "" {
-			mountPath = SandboxSkillsMountPath
-		}
-		mounts[mountPath] = strings.TrimSpace(skillMount.VolumeName)
-	}
-	return mounts
-}
-
-func skillMountFromWork(work *db.EnvironmentWork) (*SkillMount, bool) {
-	if work == nil || len(work.Metadata) == 0 || strings.TrimSpace(string(work.Metadata)) == "null" {
-		return nil, false
-	}
-	var metadata map[string]json.RawMessage
-	if err := json.Unmarshal(work.Metadata, &metadata); err != nil {
-		return nil, false
-	}
-	raw, ok := metadata[SkillMountMetadataKey]
-	if !ok || len(raw) == 0 || strings.TrimSpace(string(raw)) == "null" {
-		return nil, false
-	}
-	var mount SkillMount
-	if err := json.Unmarshal(raw, &mount); err != nil {
-		return nil, false
-	}
-	if strings.TrimSpace(mount.VolumeName) == "" {
-		return nil, false
-	}
-	return &mount, true
-}
-
-func skillMountVolumeName(manifestSHA256 string) string {
-	sha := strings.TrimSpace(manifestSHA256)
-	if sha == "" {
-		sha = "unknown"
-	}
-	return skillMountVolumePrefix + sha
 }
