+++ scripts/go-file-line-budgets.txt
@@ -28,7 +28,6 @@
 595 internal/codesessions/service.go
 571 internal/db/console_api_keys.go
 552 internal/db/deployments.go
-547 internal/runtime/e2bruntime/runtime.go
 546 internal/db/workbench.go
 537 internal/environments/environment_manager.go
 534 internal/files/platform.go
